pyproject.toml
[project]
name = "adk-porker"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "flet[all]>=0.28.3",
    "google-adk>=1.5.0",
    "python-dotenv>=1.1.1",
    "requests>=2.32.0",
]

[dependency-groups]
dev = [
    "coverage>=7.10.1",
    "pytest>=8.4.1",
]
